refactor(auth): auth.ts を責務ごとに分割 (SPR-168 Low)
204行に同居していた betterAuth コールバック実体を責務別モジュールへ抽出し、
auth.ts は betterAuth 設定の配線（薄いコールバック）に専念させる。

- guild-sync.ts: 日次 guild 同期（refreshGuildStatusIfNeeded）
- on-first-signup.ts: 初回 Discord 連携時のアプリユーザー作成（createAppUserOnFirstSignup）
- enrich-session.ts: customSession の appUser 構築（buildAppUserForSession）
- auth.ts: 上記を呼ぶ薄い customSession / account.create.after に

挙動はコード verbatim 移動で不変。型推論（customSession の appUser）も維持され
cast レスのまま。これら3モジュールは旧 auth.ts と同じ better-auth glue（Firestore 副作用主体）
のため coverage 除外に追加し計上を中立化。
